Update ReadMe and Makefile to create data
The code, data, and rawData ReadMe's now explain how where to get the data, how
to organize it, and the difference between passage and waves.

The Makefile now has dependencies on the household data for the TeX file (will
be changed later) and dependencies for the household data, together with
a recipe to creat them.