Fix treatment of education for respondent and spouse
The "grd" variable misses if you have never attended school, so new variables
"educ_years" and "sp_educ_years" are introduced. Both are equal to 0 (zero) if
never attended school. The same value is assigned if participated in adult
education or had ******* as recorded outcome. This still leave 7 with missing
education.

Also fixed that spouses were given 0 when they were supposed to be missing.