Add check on education for both respondent and spouse
Use the mode to check if education in years changes across waves. Unlike for education
there are no good years to check which is the "correct" education level. Also checks
whether education vary across years, so that which education group a
respondent ends up in might change.

Creates two education group variables

- educ017: none, some primary, graduated primary or above
- educ07: none and some primary combined, graduated primary or above

These are done for both respondent and spouse.