Add asset variables
Adds the first wave variables for assets per capita, land value per
capita, and land area for all waves.

This is based on the first wave that an individual is observed in the
survey. Hence, it is possible that different respondents within the
same household have different first wave values for these variables.