Remove recoding of missing pregnancy
Some women who reported never being pregnant were missing information on
contraception and pregnancy. This appears to be a symptom of poor coding, so
instead of trying to figure out what the "correct" coding should be these
are now left "as is". This allow them to be dropped when running the regressions.
There are only 2 observations affected by this.