Add first minimally working version of main analysis file
This mainly introduces the loading of the base data and dropping of observations
that have missing values in any of the important variables (pregnancy, birth,
contraceptive use, and crop loss).

A basic set of regression (fixed effects at woman level) for the five main
outcomes are also included. These show results that are nearly identical to
those from the first (undocumented) version of the base data set.