Change lower age cut-off to less than 17
This adds three observations, but does not materially change the
results. Given that there are only three observations it appears
easier to simply say that we include all women less than 45 who
are married/partnered and observed in all waves (there are other
young women who are dropped because of missing information/waves).

Also includes a minor edit of dropping because of missing information.
Each condition is now a separate line rather than all in one line.
This makes it easier to see which missing information is responsible
for dropping of observations.