Allow easy changing of asset and crop loss divisor
Prior assets and crop loss were always divided by the same number
(implemented to ease readability of regression results), but it
was spread over two different files. Moved code from womenCommon.do
to crBase.do, to make it possible to change only one place instead.

Also split so that assets and crop loss are now divided by different
numbers (1,000,000 for assets and 10,000 for crop loss).

All of this is in response to R1's comment on coefficient and standard
errors that showed as 0.000.

One remaining issue is that some tables have hard-coded levels for
crop loss and assets in table notes. That needs to be changed.