Update files to allow for rainfall data
New file is crRainfall_waves.do, which creates four files with
rainfall data. This requires files that has information on
district and clusters. Because the rainfall data is not in
the directory where downloaded ReadMe.md for rawData is also
updated. Makefile updated with new dependencies for rainfall
data and targets.

In addition, wave 3 data files were in lowercase format, which
led Makefile to keep running crBase.do and not being able to
find targets for wave 3 files. All file names for waves
changed to uppercase. ReadMe.md for rawData updated with code
for making files uppercase.