Add response to R1's question on abstinence
Question was whether abstinence could be recorded as a contraceptive
strategy even if it was caused by other factors, such as stress.
Added table to contraceptive effectiveness file that subdivides
contraceptive use so abstinence only is a separate category,
and updated Makefile and main paper with this new appendix table.

Fixed the definitions in exclude abstinence file.

Wrote initial response based on new results to argue that if
stress could cause this effect we should see something similar
with illness but there is no evidence of that.