Respond to R1's question on wave 1 assets
The discussion of assets interacted with crop loss was made a separate
section prior. This revision adds log of crop loss interacted with
log of assets and results where the first wave crop loss is dropped
because it might affect wave 1 assets.

The new section is edited to reflect changes in results.

All results are combined into one table, so three original analysis
files are deleted because their results show up elsewhere.